Fix #1289
