+++ libraries/chain/include/eosio/chain/types.hpp
@@ -16,6 +16,7 @@
 #include <fc/io/varint.hpp>
 #include <fc/io/enum_type.hpp>
 #include <fc/crypto/sha224.hpp>
+#include <fc/variant.hpp>
 #include <fc/optional.hpp>
 #include <fc/safe.hpp>
 #include <fc/container/flat.hpp>
@@ -234,6 +235,37 @@ namespace eosio { namespace chain {
    typedef vector<std::pair<uint16_t,vector<char>>> extensions_type;
 
 
+   template<typename T>
+   struct may_not_exist {
+      T value{};
+   };
 } }  // eosio::chain
 
 FC_REFLECT( eosio::chain::void_t, )
+
+namespace fc {
+
+template<typename ST, typename T>
+datastream<ST>& operator << (datastream<ST>& s, const eosio::chain::may_not_exist<T>& v) {
+   raw::pack(s, v.value);
+   return s;
+}
+
+template<typename ST, typename T>
+datastream<ST>& operator >> (datastream<ST>& s, eosio::chain::may_not_exist<T>& v) {
+   if (s.remaining())
+      raw::unpack(s, v.value);
+   return s;
+}
+
+template<typename T>
+void to_variant(const eosio::chain::may_not_exist<T>& e, fc::variant& v) {
+   to_variant( e.value, v);
+}
+
+template<typename T>
+void from_variant(const fc::variant& v, eosio::chain::may_not_exist<T>& e) {
+   from_variant( v, e.value );
+}
+
+} // namespace fc
